+++ src/include/gsheets_requests.hpp
@@ -13,4 +13,5 @@ enum class HttpMethod {
 std::string perform_https_request(const std::string& host, const std::string& path, const std::string& token, 
                                   HttpMethod method = HttpMethod::GET, const std::string& body = "");
 
+std::string fetch_sheet_data(const std::string& sheet_id, const std::string& token, const std::string& sheet_name);
 }
